Fix arm

+++ minidumper-test/src/lib.rs
@@ -394,7 +394,7 @@ pub fn assert_minidump(md_buf: &[u8], signal: Signal) {
                     errors::ExceptionCodeWindowsAccessType::WRITE
                 ));
 
-                assert_eq!(crash_address, sadness_generator::SEGFAULT_ADDRESS);
+                assert_eq!(crash_address, sadness_generator::SEGFAULT_ADDRESS as u64);
             }
             Signal::StackOverflow | Signal::StackOverflowCThread => {
                 verify!(CrashReason::WindowsGeneral(
@@ -477,7 +477,7 @@ pub fn assert_minidump(md_buf: &[u8], signal: Signal) {
                     errors::ExceptionCodeMacBadAccessKernType::KERN_INVALID_ADDRESS,
                 ));
 
-                assert_eq!(crash_address, sadness_generator::SEGFAULT_ADDRESS);
+                assert_eq!(crash_address, sadness_generator::SEGFAULT_ADDRESS as u64);
             }
             Signal::StackOverflow | Signal::StackOverflowCThread => {
                 verify!(CrashReason::MacBadAccessKern(
